# backend/docs/OPENAPI_SCHEMA.yaml
openapi: 3.0.3
info:
  title: Media Center Arti Marziali API
  description: |
    REST API for the Media Center Arti Marziali platform.

    ## Authentication
    Most endpoints require a Bearer token in the Authorization header.

    ## User Tiers
    - `free` - Basic access with ads
    - `hybrid_light` - No ads, 1080p
    - `hybrid_standard` - Downloads enabled
    - `premium` - 4K, all features
    - `business` - Bulk licenses, API access
  version: 1.0.0
  contact:
    name: API Support
    email: api@example.com
servers:
  - url: https://api.example.com/api/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
    UserRegisterRequest:
      type: object
      required:
        - email
        - password
        - username
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        username:
          type: string
          minLength: 3
          maxLength: 50
        full_name:
          type: string
    UserLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          enum: [bearer]
        expires_in:
          type: integer
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        full_name:
          type: string
        tier:
          type: string
          enum: [free, hybrid_light, hybrid_standard, premium, business]
        is_active:
          type: boolean
        is_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    BlockchainBatchResponse:
      type: object
      properties:
        batch_id:
          type: string
          format: uuid
        batch_date:
          type: string
          format: date-time
        period_start:
          type: string
          format: date-time
        period_end:
          type: string
          format: date-time
        status:
          type: string
        total_views:
          type: integer
        unique_users:
          type: integer
        total_watch_time:
          type: integer
        total_revenue:
          type: number
        data_hash:
          type: string
        blockchain_tx_hash:
          type: string
        explorer_url:
          type: string
...
    Error:
      type: object
      properties:
        detail:
          type: string
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Register new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already registered
  /auth/login:
    post:
      tags: [auth]
      summary: Login user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
  /auth/me:
    get:
      tags: [auth]
      summary: Get current user
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /auth/logout:
    post:
      tags: [auth]
      summary: Logout user
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logged out successfully
...
  /blockchain/batches/create:
    post:
      tags: [blockchain]
      summary: Create weekly batch (admin)
      operationId: createBlockchainBatch
      security:
        - BearerAuth: []
      parameters:
        - name: week_offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Batch created
  /blockchain/batches/{batch_id}:
    get:
      tags: [blockchain]
      summary: Get batch status
      operationId: getBatchStatus
      parameters:
        - name: batch_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Batch details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockchainBatchResponse'
  /blockchain/batches/{batch_id}/publish:
    post:
      tags: [blockchain]
      summary: Publish batch to blockchain (admin)
      operationId: publishBatch
      security:
        - BearerAuth: []
      parameters:
        - name: batch_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Batch published
...